+++ src/components/home/ChallengeList.jsx
@@ -1,28 +1,12 @@
-import React, { useEffect, useState } from "react";
-import axios from "axios";
+import React from "react";
 
 import { Grid } from "@mui/material";
 import ChallengeCard from "./ChallengeCard";
 
-const ChallengeList = () => {
-	const [list, setList] = useState([]);
-
-	const fetchChallenges = async () => {
-		try {
-			const { data } = await axios.get("/challenges?kind=All");
-			setList(data);
-		} catch (e) {
-			console.error(e);
-		}
-	};
-
-	useEffect(() => {
-		fetchChallenges();
-	}, []);
-
+const ChallengeList = ({ ...props }) => {
 	return (
 		<Grid container spacing={4}>
-			{list.map((challenge) => (
+			{props.list.map((challenge) => (
 				<Grid item xs={4} key={challenge.challengeId}>
 					<ChallengeCard
 						key={challenge.challengeId}
@@ -30,8 +14,8 @@ const ChallengeList = () => {
 						title={challenge.title}
 						category={challenge.category.name}
 						locationType={challenge.locationType}
-						startDate={challenge.startDate}
-						endDate={challenge.endDate}
+						startDate={challenge.challengeStartDate}
+						endDate={challenge.challengeEndDate}
 						imgUrl={challenge.imgUrl}
 						limitPerson={challenge.limitPerson}
 						participants={challenge.participants}
